requirements.txt
arrow==0.8.0
binaryornot==0.4.0
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
Django==1.10.1
django-betterforms==1.1.4
django-crispy-forms==1.6.1
django-malwareanalystreporter==1.0
fpdf==1.7.2
future==0.16.0
Jinja2==2.8
jinja2-time==0.2.0
Markdown==2.4.1
MarkupSafe==0.23
Pillow==3.4.2
poyo==0.4.0
psycopg2==2.6.2
PyPDF2==1.26.0
python-dateutil==2.6.0
reportlab==3.3.0
six==1.10.0
tagging==0.2.1
virtualenv==15.0.3
whichcraft==0.4.0
